requirements.txt
streamlit
numpy
pypdf
sentence-transformers
rank-bm25
faiss-cpu
transformers